Support creating multiple models in one query. Fixes #12.
Introduces a `Model::create_many()` method and `Simple_Query::insert_many()`.
Will prefer to do multiple inserts in one query. Can be configured to just
wrap `Simple_Query::insert()`. This would be necessary for innodb configured
to use interleaving. In this case, transactions will be used to ensure it is
all or nothing.