Allow for InMemoryTable to accept callable column definitions.
The register_table() helper can be used for defining foreign columns
or foreign models. These require the table to be accessible to construct
the column. However, if the other table hasn't registered yet, or the
table is self-referential, it makes it impossible for those to be defined.
Lazy loading columns allows us to wait until all the tables have been most
likely registered.